recipe/item.go: Do not get item from block state
Since the Bedrock Edition Flattening has finished, each block has its
own unique ID and a meta value of 0. There are a few exceptions, such as
beds and banners, where the meta values do not map to different block
states but rather to block entity NBT. Therefore, there is no longer a
need to use the block state and we should stop using it since it is
causing problems.